Fix typo in VMap BIH generation
Due to this typo some objects could be left out during BIH generation and thus never checked for intersection.
This breaks height and LoS calculations in some areas.

This occurrence of this bug can be seen in "Md_Mushroomcave.wmo". It can be found in several locations,
for example (-6140, -2960, 400), map 0.
In some parts of the cave mobs will either fall through the ground or fly up into the sky when chasing the player.
LoS is not obstructed there as well.

P.S. This typo presents in all MaNGOS cores.